refactor(home): valuation hero renders inside ChatGreeting — no chat.tsx changes, no prop drilling (KISS/OCP)
Review feedback on #1852: chat.tsx should not be modified in this PR, and
header composition via props is drilling context that providers already
own. ChatGreeting — the component chat.tsx already delegates its
empty-state header to — becomes the artist header: the artist-scoped
valuation hero when measured data exists for the current scope (via
useHomeValuation, provider-backed), otherwise the greeting plus a 'Get a
free catalog valuation' CTA into the funnel. chat.tsx, NewChatBootstrap
and HomePage are byte-identical to main in this PR; every empty-chat
surface gets the same single flow for free.
